Created classes for each separate tab
Not to be confused with the JavaFX tabs, these tab objects initialize everything from the visuals to the functions of a tab
