+++ src/sample/Tabs.java
@@ -1,0 +1,7 @@
+package sample;
+
+import javafx.scene.control.Tab;
+
+public interface Tabs {
+    Tab init();
+}
